Update redshift-splitter.py
Updated and added new features

+++ redshift-splitter.py
@@ -1,10 +1,4 @@
-import os
-import json
-import boto3
-import gzip
-import argparse
-import shutil
-
+import os, json, boto3, gzip, argparse, shutil
 from datetime import datetime
 
 # Handle Arguments
@@ -12,9 +6,12 @@ parser = argparse.ArgumentParser(description='Watch an SQS Queue and process red
 parser.add_argument('-q','--queue', help='Name of the Queue to watch', required=True)
 parser.add_argument('-d','--destination', help='Destination bucket', required=True)
 parser.add_argument('-i','--input', help='Input File Type', default='gz')
-parser.add_argument('-c','--chunk', help='Chunk Size to split file', default='100M')
+parser.add_argument('-H','--headers', help='Include headers in each file - Redshift does not automatically provide them', action='store_true')
+parser.add_argument('-HL','--header-list', help='A quoted list of headers in the same order and delimeter as your data (this wont be validated)', required=False)
+parser.add_argument('-c','--chunk', help='Chunk Size to split file (in Megabytes)', default='200')
 parser.add_argument('-r','--region', help='AWS Region of the bucket', default='us-west-2')
 parser.add_argument('-f','--filename', help='Filename Prefix', default='redshift')
+parser.add_argument('-k','--keep-files', help='Keep all temporary files (Space Warning)', action='store_true')
 parser.add_argument('-z','--zip', help='Compression output to use (uses <command> * so must support this)', default='bzip2')
 args = vars(parser.parse_args())
 
@@ -60,8 +57,8 @@ def sync_to_s3(target_dir, aws_region=args['region'], bucket_name=args['destinat
 	for args['filename'] in os.listdir(target_dir):
 		s3.Object(bucket_name, folder + '/' + args['filename']).put(Body=open(os.path.join(target_dir, args['filename']), 'rb'))
 
-		print('File uploaded to https://s3.%s.amazonaws.com/%s/%s' % (
-			args['region'], bucket_name, args['filename']))
+		print('File uploaded to https://s3.%s.amazonaws.com/%s/%s/%s' % (
+			args['region'], bucket_name, redshift_folder, args['filename']))
 
 # Print queue for debugging
 print(queue.url)
@@ -73,10 +70,12 @@ for message in queue.receive_messages(MaxNumberOfMessages=1,WaitTimeSeconds=10):
 
 	bucket = js['Records'][0]['s3']['bucket']['name']
 	key = js['Records'][0]['s3']['object']['key']
+	size = str(js['Records'][0]['s3']['object']['size'])
 
 	print("= Details from SQS =")
 	print("Bucket: " + bucket)
 	print("File: " + key)
+	print("Size: " + size)
 
 	if not os.path.exists(folder):
 		os.makedirs(folder)
@@ -98,8 +97,33 @@ for message in queue.receive_messages(MaxNumberOfMessages=1,WaitTimeSeconds=10):
 	
 	print("= Splitting file into chunks =")
 	
-	# Split into 100MB chunks
-	os.system('split -d -b ' + args['chunk'] + ' ' + args['filename'] + ' ' + args['filename'] + '-part-')
+	# Split into chunks
+	numchunks = int(size) / (int(args['chunk']) * 1000000) + 1
+	numchunks= str(numchunks)
+
+	# Force header addition to each chunk
+	if args['header_list']:
+		os.system('echo "%s" > tmp_file' % args['header_list'])
+		os.system('cat '+ args['filename'] + ' >> tmp_file')
+		os.system('mv -f tmp_file '+ args['filename'])
+
+	os.system('split -e -d --number=l/' + numchunks + ' ' + args['filename'] + ' ' + args['filename'] + '-part-')
+
+	if args['headers']:
+		print("= Adding Headers to Splits =")
+		cmd = '''
+		for file in %s-part-*;  do
+			if [ "$file" !=  "%s-part-00" ] ;
+				then
+				    head -n 1 %s > tmp_file
+				    cat $file >> tmp_file
+				    mv -f tmp_file $file
+			fi
+		done
+		''' % (args['filename'], args['filename'],args['filename'])
+		
+		os.system(cmd)
+
 
 	print("= Rezipping files =")
 	
@@ -112,7 +136,7 @@ for message in queue.receive_messages(MaxNumberOfMessages=1,WaitTimeSeconds=10):
 	print("= Deleting source files =")
 
 	# Delete original file
-	os.remove(args['filename'])
+	#os.remove(args['filename'])
 
 	# Re-upload
 
@@ -132,11 +156,16 @@ for message in queue.receive_messages(MaxNumberOfMessages=1,WaitTimeSeconds=10):
 	# Upload the files to 
 	sync_to_s3(folder, args['region'], args['destination'], redshift_folder)
 
-	print("= Cleaning up =")
 
-	shutil.rmtree(folder)
+	if args['keep_files']:
+		# Clean up if requested
+		print("= Cleaning up =")
+
+		shutil.rmtree(folder)
+
 	os.listdir('.')
 
+	# Tell SQS we are done with the file
 	print("= Marking complete in SQS =")
 
 	message.delete()
